+++ sdk/src/androidTest/java/ly/count/android/sdk/scUP_UserProfileTests.java
@@ -94,7 +94,7 @@ public class scUP_UserProfileTests {
     /**
      * Related user properties should be saved before event recordings
      * call order, user property with "dark_mode", event, user property with "light_mode"
-     * generated request order first user property request + 3 events + user property request with light_mode
+     * generated request order first user property request + 3 events + user property request with light_mode + begin session
      */
     @Test
     public void eventSaveScenario_changeDeviceIDWithoutMerge() throws JSONException {
@@ -111,11 +111,11 @@ public class scUP_UserProfileTests {
         countly.userProfile().setProperty("theme", "light_mode");
         TestUtils.assertRQSize(1); // no request is generated on the way
 
-        countly.deviceId().changeWithoutMerge("new_device_id");
+        countly.deviceId().changeWithoutMerge("new_device_id"); // this will begin a new session
 
         // first user property request + 3 events + user property request with light_mode
-        ModuleUserProfileTests.validateUserProfileRequest(0, 3, TestUtils.map(), TestUtils.map("theme", "dark_mode"));
-        ModuleUserProfileTests.validateUserProfileRequest(2, 3, TestUtils.map(), TestUtils.map("theme", "light_mode"));
+        ModuleUserProfileTests.validateUserProfileRequest(0, 4, TestUtils.map(), TestUtils.map("theme", "dark_mode"));
+        ModuleUserProfileTests.validateUserProfileRequest(2, 4, TestUtils.map(), TestUtils.map("theme", "light_mode"));
     }
 
     /**