+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+## XX.XX.XX
+* Mitigated issues where:
+  * session was ending regardless of manual control after without merge, not anymore.
+  * session was not starting even if consent is not required and automatic sessions are enabled after without merge, not anymore.
+
 ## 24.7.0
 * ! Minor breaking change ! User properties will now be automatically saved under the following conditions:
   * When an event is recorded